feat: refactored tests - moved ballot tests to integration, created separate mock data module
