+++ src/lib.rs
@@ -3,4 +3,6 @@ pub mod error;
 pub mod leader;
 pub mod message;
 pub mod party;
+#[cfg(any(test, feature = "test-mocks"))]
+pub mod test_mocks;
 pub mod value;
